requirements.txt
urllib3==1.24.1
lxml==4.2.5
beautifulsoup4==4.6.3
numpy==1.15.4
pandas==0.23.4
python-dotenv==0.10.1